# contracts/openapi.json
{
  "components": {
    "schemas": {
      "BatchSummaryOut": {
        "additionalProperties": false,
        "description": "The three numbers, plus the correctness figures that qualify them.",
        "properties": {
          "at_risk": {
            "$ref": "#/components/schemas/Money"
          },
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "exceptions": {
            "title": "Exceptions",
            "type": "integer"
          },
          "exceptions_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Exceptions By Type",
            "type": "object"
          },
          "false_matches": {
            "description": "Wrong auto-matches. Reported separately from match rate on purpose \u2014 blending them hides the failure that costs money.",
            "title": "False Matches",
            "type": "integer"
          },
          "match_rate": {
            "title": "Match Rate",
            "type": "number"
          },
          "matched": {
            "title": "Matched",
            "type": "integer"
          },
          "records_accepted": {
            "title": "Records Accepted",
            "type": "integer"
          },
          "records_deduplicated": {
            "title": "Records Deduplicated",
            "type": "integer"
          },
          "records_in": {
            "title": "Records In",
            "type": "integer"
          },
          "records_rejected": {
            "title": "Records Rejected",
            "type": "integer"
          },
          "throughput_per_second": {
            "title": "Throughput Per Second",
            "type": "number"
          },
          "unexplained": {
            "description": "Records neither matched nor excepted. Must be 0.",
            "title": "Unexplained",
            "type": "integer"
          }
        },
        "required": [
          "batch_id",
          "records_in",
          "records_accepted",
          "records_deduplicated",
          "records_rejected",
          "matched",
          "match_rate",
          "throughput_per_second",
          "exceptions",
          "at_risk",
          "false_matches",
          "unexplained",
          "exceptions_by_type"
        ],
        "title": "BatchSummaryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Money": {
        "additionalProperties": false,
        "description": "An amount and its currency, travelling together.\n\nBundled rather than sent as loose fields so an amount cannot be read without\nits currency. A bare number is the bug that turns 40,000 USD into 40,000 INR.",
        "properties": {
          "amount_minor": {
            "description": "Amount in integer minor units (paise for INR). NEVER a float and never pre-formatted: pair it with `currency` and format at the display edge.",
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount_minor",
          "currency"
        ],
        "title": "Money",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "An autonomous finance controller. Reconciles transactions across gateway, settlement, bank and internal ledger, and reports every record it refuses to guess at as a typed exception with the money at risk.\n\n**Money is always integer minor units** (paise for INR) paired with a currency \u2014 never a float, never pre-formatted. Format at the display edge.\n\nEvery endpoint requires a bearer token. Collections are cursor-paginated.",
    "title": "Ledger",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/batches/current": {
      "get": {
        "operationId": "current_batch_v1_batches_current_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summary of the current batch",
        "tags": [
          "batches"
        ]
      }
    },
...
  }
}